Firebase showing up in react-native example.
