+++ examples/complete/react-native/package.json
@@ -10,9 +10,13 @@
   "dependencies": {
     "react": "15.4.1",
     "react-native": "0.42.0",
-    "react-native-google-signin": "0.9.0"
+    "react-native-google-signin": "0.9.0",
+    "react-redux": "^5.0.3",
+    "react-redux-firebase": "^1.4.0-alpha",
+    "redux": "^3.6.0"
   },
   "devDependencies": {
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-react-native": "1.9.1",
     "concurrently": "^3.4.0",
     "fs-extra": "^2.0.0",
