+++ examples/complete/react-native/src/reducer.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import { firebaseStateReducer as firebase } from 'react-redux-firebase'
+
+const rootReducer = combineReducers({
+  firebase
+})
+
+export default rootReducer
