+++ examples/complete/react-native/src/store.js
@@ -1,0 +1,28 @@
+import { createStore, compose } from 'redux'
+import rootReducer from './reducer'
+import { firebase as fbConfig } from './config'
+import { reactReduxFirebase } from 'react-redux-firebase'
+
+export default function configureStore (initialState, history) {
+  const createStoreWithMiddleware = compose(
+    reactReduxFirebase(fbConfig,
+      {
+        userProfile: 'users',
+        enableLogging: false,
+        enableRedirectHandling: false
+      }
+    ),
+    typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? window.devToolsExtension() : f => f
+  )(createStore)
+  const store = createStoreWithMiddleware(rootReducer)
+
+  if (module.hot) {
+    // Enable Webpack hot module replacement for reducers
+    module.hot.accept('./reducer', () => {
+      const nextRootReducer = require('./reducer')
+      store.replaceReducer(nextRootReducer)
+    })
+  }
+
+  return store
+}
